+++ src/commands/stash/gitPop.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import * as vscode from 'vscode';
-import * as moment from 'moment';
+import * as moment from 'moment'; 
 import strings from '../../constants/string-constnats';
 import {
     exec
@@ -9,7 +9,7 @@ import {
 import * as logger from "../../logger";
 
 export function activate(context: vscode.ExtensionContext) {
-    let disposable = vscode.commands.registerCommand('extension.gitUnstash', () => {
+    let disposable = vscode.commands.registerCommand('extension.gitPop', () => {
         exec(strings.git.stash("list "), {
             cwd: vscode.workspace.rootPath
         }, (error, stdout, stderr) => {
@@ -49,15 +49,17 @@ export function activate(context: vscode.ExtensionContext) {
                 }
                 return stashObject;
             });
-            vscode.window.showQuickPick(stashList, {matchOnDescription: true, placeHolder: "Choose what to unstash"}).then(chosenitem => {
+            vscode.window.showQuickPick(stashList, {matchOnDescription: true, placeHolder: "Choose the stash you wish to apply and drop"}).then(chosenitem => {
                 if(chosenitem === undefined){return}
                 let stashIndex = chosenitem.description.substring(chosenitem.description.indexOf("(")+1, chosenitem.description.indexOf(")"));
-                exec(strings.git.stash("apply " + stashIndex), { cwd: vscode.workspace.rootPath}, (error, stdout, stderr) => {
+                exec(strings.git.stash("pop " + stashIndex), { cwd: vscode.workspace.rootPath}, (error, stdout, stderr) => {
                     if(error) {
-                        logger.logError(strings.error("unstashing:"), stderr || error);
+                        logger.logError(strings.error("droping stash:"), stderr || error);
                         return;
                     }
-                    logger.logInfo("Stash was applied on current branch");
+                    if(stdout.indexOf("Dropped") != -1){
+                        logger.logInfo(strings.success.general("Stash", "applied and removed"));
+                    }
                 });
             });
         });