+++ src/logger.ts
@@ -14,21 +14,21 @@ import strings from './constants/string-constnats';
  * The output channel instance
  * @type {vscode.OutputChannel}
  */
-let outLogChannel: vscode.OutputChannel;
+let outLogChannel = {};
 
 /**
  * Creates an output channel instance or retrives it if already exists.
  * @returns {vscode.OutputChannel} The output channel instance
  */
-function getLogChannel() {
-    if (outLogChannel === undefined) {
-        outLogChannel = vscode.window.createOutputChannel('Git Merger');
+function getLogChannel(type) {
+    if (outLogChannel[type] === undefined) {
+        outLogChannel[type] = vscode.window.createOutputChannel('Git merger ' + type);
     }
-    return outLogChannel;
+    return outLogChannel[type];
 }
 
 export function logError(errorTitle: string, error: any) {
-    getLogChannel().appendLine(`[Error-${moment().format(strings.timeForamt.hours)}] ${errorTitle}\n${error.toString()}`);
+    getLogChannel("errors").appendLine(`[Error-${moment().format(strings.timeForamt.hours)}] ${errorTitle}\n${error.toString()}`);
     vscode.window.showErrorMessage(strings.windowErrorMessage, strings.actionButtons.openLog).then((action) => {
         if (action == strings.actionButtons.openLog) {
             this.openLog();
@@ -37,18 +37,18 @@ export function logError(errorTitle: string, error: any) {
 }
 
 export function logInfo(message: string) {
-    getLogChannel().appendLine(`[Info-${moment().format(strings.timeForamt.hours)}] ${message}`);
+    getLogChannel("info").appendLine(`[Info-${moment().format(strings.timeForamt.hours)}] ${message}`);
     vscode.window.showInformationMessage(message);
 }
 
 export function logWarning(message: string) {
-    getLogChannel().appendLine(`[Warning-${moment().format(strings.timeForamt.hours)}] ${message}`);
+    getLogChannel("error").appendLine(`[Warning-${moment().format(strings.timeForamt.hours)}] ${message}`);
 }
 
 export function logDebug(message: string) {
-    getLogChannel().appendLine(`[Debug-${moment().format(strings.timeForamt.hours)}] ${message}`);
+    getLogChannel("info").appendLine(`[Debug-${moment().format(strings.timeForamt.hours)}] ${message}`);
 }
 
-export function openLog(): void {
-    getLogChannel().show();
+export function openLog(type): void {
+    getLogChannel(type).show();
 }